refactor(eventrouter): remove legacy routing-rule translator
Routing rules are conditions-native. The translate-on-read path that
converted pre-conditions flat-matcher rows into conditions on decode is
now dead weight:

- Drop model.LegacyRoutingRule and its HasMatcher helper.
- Delete eventrouter.TranslateRule and the emitsAll fan-out helper.
- Remove store.RuleTranslator and the store.Store.Rules field, along
  with the server startup wiring that injected DefaultSchemaRegistry as
  the translator.
- Simplify DecodeRoutingRules to a plain json.Unmarshal (now a
  package-level function, since it no longer depends on the store) and
  drop the "store.Rules is nil" error path.

SchemaRegistry.Validate and SchemaRegistry.DefaultRules (rule validation
and the ruleless-org default fallback) are unaffected.

Add a migration that normalizes any surviving legacy flat-matcher rows
in orgs.routing_rules to the conditions shape, so they don't silently
lose their matcher on decode.
